[RUST] Rename structural common module
