+++ rust/tvm-ffi/src/extra/mod.rs
@@ -16,8 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-mod common;
 pub mod dispatch;
 pub mod module;
+mod structural_common;
 pub mod structural_mutate;
 pub mod structural_visit;
