+++ rust/tvm-ffi/src/extra/structural_common.rs
